+++ src/python/pants/backend/experimental/java/lint/google_java_format/register.py
@@ -5,7 +5,7 @@ from pants.backend.java.lint.google_java_format import rules as fmt_rules
 from pants.backend.java.lint.google_java_format import skip_field
 from pants.jvm import jdk_rules
 from pants.jvm import util_rules as jvm_util_rules
-from pants.jvm.resolve import coursier_fetch, jvm_tool
+from pants.jvm.resolve import jvm_tool, user_resolves
 
 
 def rules():
@@ -14,6 +14,6 @@ def rules():
         *skip_field.rules(),
         *jdk_rules.rules(),
         *jvm_tool.rules(),
-        *coursier_fetch.rules(),
+        *user_resolves.rules(),
         *jvm_util_rules.rules(),
     ]
