+++ src/python/pants/backend/java/dependency_inference/java_parser.py
@@ -19,7 +19,7 @@ from pants.engine.fs import AddPrefix, Digest, DigestContents
 from pants.engine.process import BashBinary, FallibleProcessResult, Process, ProcessExecutionFailure
 from pants.engine.rules import Get, MultiGet, collect_rules, rule
 from pants.jvm.jdk_rules import JdkSetup
-from pants.jvm.resolve.coursier_fetch import MaterializedClasspath, MaterializedClasspathRequest
+from pants.jvm.resolve.user_resolves import MaterializedClasspath, MaterializedClasspathRequest
 from pants.option.global_options import ProcessCleanupOption
 from pants.util.logging import LogLevel
