+++ src/python/pants/backend/scala/compile/scalac_plugins.py
@@ -20,9 +20,9 @@ from pants.engine.unions import UnionRule
 from pants.jvm.goals import lockfile
 from pants.jvm.goals.lockfile import GenerateJvmLockfile
 from pants.jvm.resolve.common import ArtifactRequirements, CoursierResolvedLockfile
-from pants.jvm.resolve.coursier_fetch import MaterializedClasspath, MaterializedClasspathRequest
 from pants.jvm.resolve.jvm_tool import GatherJvmCoordinatesRequest
 from pants.jvm.resolve.jvm_tool import rules as jvm_tool_rules
+from pants.jvm.resolve.user_resolves import MaterializedClasspath, MaterializedClasspathRequest
 from pants.util.ordered_set import FrozenOrderedSet
 from pants.util.strutil import bullet_list
