+++ src/python/pants/jvm/compile_test.py
@@ -47,9 +47,9 @@ from pants.jvm.compile import (
 )
 from pants.jvm.goals import lockfile
 from pants.jvm.resolve.common import CoursierResolveKey
-from pants.jvm.resolve.coursier_fetch import CoursierFetchRequest
-from pants.jvm.resolve.coursier_fetch import rules as coursier_fetch_rules
 from pants.jvm.resolve.coursier_setup import rules as coursier_setup_rules
+from pants.jvm.resolve.user_resolves import CoursierFetchRequest
+from pants.jvm.resolve.user_resolves import rules as coursier_fetch_rules
 from pants.jvm.target_types import JvmArtifactTarget
 from pants.jvm.testutil import (
     RenderedClasspath,
