+++ src/python/pants/jvm/jdk_rules_test.py
@@ -14,8 +14,8 @@ from pants.engine.process import BashBinary, Process, ProcessResult
 from pants.engine.process import rules as process_rules
 from pants.jvm.jdk_rules import JdkSetup, parse_jre_major_version
 from pants.jvm.jdk_rules import rules as jdk_rules
-from pants.jvm.resolve.coursier_fetch import rules as coursier_fetch_rules
 from pants.jvm.resolve.coursier_setup import rules as coursier_setup_rules
+from pants.jvm.resolve.user_resolves import rules as coursier_fetch_rules
 from pants.jvm.testutil import maybe_skip_jdk_test
 from pants.jvm.util_rules import rules as util_rules
 from pants.testutil.rule_runner import PYTHON_BOOTSTRAP_ENV, QueryRule, RuleRunner
