+++ src/python/pants/jvm/resolve/jvm_tool_test.py
@@ -14,9 +14,9 @@ from pants.jvm.goals.lockfile import GenerateJvmLockfile, GenerateJvmLockfileFro
 from pants.jvm.goals.lockfile import rules as lockfile_rules
 from pants.jvm.resolve import jvm_tool
 from pants.jvm.resolve.common import Coordinate
-from pants.jvm.resolve.coursier_fetch import rules as coursier_fetch_rules
 from pants.jvm.resolve.coursier_setup import rules as coursier_setup_rules
 from pants.jvm.resolve.jvm_tool import JvmToolBase
+from pants.jvm.resolve.user_resolves import rules as coursier_fetch_rules
 from pants.jvm.target_types import JvmArtifactTarget
 from pants.jvm.util_rules import rules as util_rules
 from pants.testutil.rule_runner import PYTHON_BOOTSTRAP_ENV, QueryRule, RuleRunner
