+++ src/python/pants/jvm/resolve/user_resolves_integration_test.py
@@ -19,7 +19,7 @@ from pants.jvm.resolve.common import (
     CoursierLockfileEntry,
     CoursierResolvedLockfile,
 )
-from pants.jvm.resolve.coursier_fetch import rules as coursier_fetch_rules
+from pants.jvm.resolve.user_resolves import rules as coursier_fetch_rules
 from pants.jvm.target_types import JvmArtifactTarget
 from pants.jvm.testutil import maybe_skip_jdk_test
 from pants.jvm.util_rules import ExtractFileDigest
