+++ src/python/pants/jvm/test/junit.py
@@ -24,8 +24,8 @@ from pants.jvm.goals import lockfile
 from pants.jvm.goals.lockfile import GenerateJvmLockfile, GenerateJvmLockfileFromTool
 from pants.jvm.jdk_rules import JdkSetup
 from pants.jvm.resolve.common import CoursierResolvedLockfile
-from pants.jvm.resolve.coursier_fetch import MaterializedClasspath, MaterializedClasspathRequest
 from pants.jvm.resolve.jvm_tool import ValidatedJvmToolLockfileRequest
+from pants.jvm.resolve.user_resolves import MaterializedClasspath, MaterializedClasspathRequest
 from pants.jvm.subsystems import JvmSubsystem
 from pants.jvm.target_types import JunitTestSourceField
 from pants.util.logging import LogLevel
